mirror

+++ srcs/get_cam_ray.cl
@@ -7,8 +7,8 @@ t_vector3d  get_cam_ray(int x, int y, t_canvas canvas, int width, int height)
 
 	x -= width / 2;
     y -= height / 2;
-	canvas_cell.x = (float)x * canvas.width / (float)width;
-    canvas_cell.y = (float)y * canvas.height / (float)height;
+	canvas_cell.x = x * canvas.width / width;
+    canvas_cell.y = y * canvas.height / height;
     canvas_cell.z = canvas.min_distance;
 	cam_ray = canvas_cell;
     return (cam_ray);
